.history/requirements_20240212082836.txt
amqp==5.1.1
asgiref==3.7.2
attrs
audio-metadata
bidict==0.22.1
billiard==3.6.4.0

cachetools==5.3.1
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.7.0
cron-descriptor==1.4.0
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.0.0
django-countries==7.5.1
django-filter==23.2
django-ipware==5.0.0
django-phonenumber-field==7.1.0
django-redis==5.2.0
django-split-settings==1.2.0
django-storages==1.13.2
django-timezone-field==5.1
djangorestframework==3.14.0
drf-api-logger==1.1.12
drf-spectacular==0.26.2
drf-yasg==1.21.7
filetype==1.2.0
furl==2.1.3
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema
jsonschema-specifications
kombu==5.3.1
more-itertools==8.14.0
orderedmultidict==1.0.1
packaging==23.1
pendulum==2.1.2
phonenumberslite==8.13.14
Pillow==9.5.0
pprintpp==0.4.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pyparsing==3.1.0
pyscopg2==66.0.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
redis==4.5.5
referencing
regex==2023.5.5
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.1
sentry-sdk==1.24.0
six==1.16.0
sqlparse==0.4.4
tinytag==1.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.4.0
wrapt==1.15.0